Remove transclusion restoring the parent level props in the keyword
When removing the transclusion, if the previous character is a
transclusion is not at the level that we want the keyword to be, the
keyword will inherit the wrong properties putting it in the wrong
transclusion. For that reason we take a principled approach to finding
the parent level of the transclusion.